// src/api/generated/getMoviesResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Cinema Platform API
 * API Gateway for Cinema Platform microservices
 * OpenAPI spec version: 1.0
 */

export interface GetMoviesResponse {
  id: string;
  title: string;
  slug?: string;
  poster: string;
  ratingAge?: number;
  releaseDate?: string;
}